backend/requirements.txt
fastapi[standard]>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy[asyncio]>=2.0.36
asyncpg>=0.30.0
alembic>=1.14.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.18
httpx>=0.28.0
google-auth>=2.37.0
google-auth-oauthlib>=1.2.1
google-api-python-client>=2.159.0
openai>=1.58.0
anthropic>=0.42.0
google-generativeai>=0.8.0
stripe>=11.4.0
python-dateutil>=2.9.0
PyPDF2>=3.0.1
pdf2image>=1.17.0
Pillow>=11.1.0
